+++ app/boards/[boardId]/styles.module.css
@@ -73,6 +73,7 @@
     border-radius: 8px;
     border: 1px 0px 0px 0px;
     opacity: 0px;
+    border: 1px solid var(--gray-B, #000000);
 }
 
 .penbutton {
@@ -82,4 +83,5 @@
     gap: 8px;
     border-radius: 8px;
     border: 1px 0px 0px 0px;
+    border: 1px solid var(--gray-B, #000000);
 }
