+++ app/globals.css
@@ -41,5 +41,5 @@ hr {
 
 img {
     margin-right: 10px;
-    border-radius: 8px;
+    /* border-radius: 8px; */
 }
