+++ app/layout.tsx
@@ -4,8 +4,8 @@ import './globals.css';
 import ApolloSetting from '../next-project/src/commons/settings/apollo-setting';
 
 export const metadata: Metadata = {
-    title: 'Create Next App',
-    description: 'Generated by create next app',
+    title: '지금 놀자',
+    description: '야놀자를 대채할 플랫폼',
 };
 
 export default function RootLayout(props: { children: React.ReactNode }) {
