+++ app/styles.module.css
@@ -1,183 +1,38 @@
 .layout {
+}
+
+.homeHeader {
     width: 1280px;
-    height: auto;
-    /* background-color: orange; */
-}
-
-.requiedmention {
-    padding: 5px 0;
-    color: red;
-    display: grid;
-    grid-template-columns: 1fr 1fr;
-}
-
-.regititle {
-    font-size: 20px;
-    font-weight: 700;
-    line-height: 28px;
-    text-align: left;
-    padding: 40px 0;
-}
-
-.writerpasswordtitle {
-    display: grid;
-    grid-template-columns: 1fr 1fr;
-}
-
-.writerpasswordinputbox {
-    display: grid;
-    grid-template-columns: 1fr 1fr;
-}
-
-.writerpasswordinput {
-    width: 620px;
-    height: 48px;
-    gap: 8px;
-    opacity: 0px;
-    border: 1px solid var(--gray-200, #d4d3d3);
-
-    background: var(--Gray-W, #ffffff);
-    border-radius: 8px;
-}
-
-.titlebox {
-}
-
-.titleinput {
-    width: 1280px;
-    height: 48px;
-    gap: 8px;
-    opacity: 0px;
-    border: 1px solid var(--gray-200, #d4d3d3);
-    border-radius: 8px;
-}
-
-.contentsbox {
-    padding: 40px 0;
-}
-
-.contentstextarea {
-    width: 1280px;
-    height: 336px;
-    gap: 8px;
-    opacity: 0px;
-    border-radius: 8px;
-    border: 1px solid var(--gray-200, #d4d3d3);
-}
-
-.addressbox {
-    padding: 40px 0;
-
-    border-bottom: 1px solid var(--gray-200, #d4d3d3);
-}
-
-.addressnumberbox {
+    height: 80px;
     display: flex;
     flex-direction: row;
-    gap: 10px;
-    padding: 8px 0;
+    justify-content: space-between;
+    align-items: center;
 }
-
-.addressnumber {
-    width: 82px;
-    height: 48px;
-    padding: 12px 16px 12px 16px;
-    gap: 0px;
-    border-radius: 8px;
-    border: 1px 0px 0px 0px;
-    opacity: 0px;
-    background: var(--Gray-W, #ffffff);
-    border: 1px solid var(--Gray-Gray-200, #d4d3d3);
-}
-
-.addressnumberbutton {
-    font-weight: 600;
-    line-height: 24px;
-    text-align: center;
-    width: 130px;
-    height: 48px;
-    padding: 12px 16px 12px 16px;
-    gap: 8px;
-    border-radius: 8px;
-    border: 1px 0px 0px 0px;
-    opacity: 0px;
-}
-
-.addressinput {
-    width: 1280px;
-    height: 48px;
-    opacity: 0px;
-    border: 1px solid var(--gray-200, #d4d3d3);
-    margin-bottom: 8px;
-    border-radius: 8px;
-}
-
-.addressdetailinput {
-    width: 1280px;
-    height: 48px;
-    opacity: 0px;
-    border: 1px solid var(--gray-200, #d4d3d3);
-    border-radius: 8px;
-}
-
-.addressinputbox {
-    display: block;
-}
-
-.youtubeBox {
-    padding: 40px 0;
-}
-
-.youtubeBoxinput {
-    width: 1280px;
-    height: 48px;
-    padding: 12px 16px 12px 16px;
-    gap: 0px;
-    border-radius: 8px;
-    border: 1px 0px 0px 0px;
-    opacity: 0px;
-    border: 1px solid var(--Gray-Gray-200, #d4d3d3);
-}
-
-.photoBox {
-    padding: 40px 0;
-}
-
-.imagebox {
+.homeHeaderLeft {
     display: flex;
     flex-direction: row;
-    gap: 8px;
+    gap: 16px;
 }
 
-.bottombuttonbox {
+.homeLoginButtonBox {
+    width: 100px;
+    height: 40px;
+    border: 1px solid #000;
     display: flex;
-    flex-direction: row;
-    justify-content: flex-end;
+    align-items: center;
+    padding: 8px 12px 8px 12px;
     gap: 8px;
-    margin-bottom: 100px;
-}
-
-.bottombuttoncancle {
-    width: 95px;
-    height: 48px;
-    padding: 12px 16px 12px 16px;
-
-    border-radius: 8px;
-    opacity: 0px;
-    border: none;
-    cursor: pointer;
-}
-
-.bottombuttonregi {
-    width: 95px;
-    height: 48px;
-    padding: 12px 16px 12px 16px;
-
-    border-radius: 8px;
-    opacity: 0px;
-    border: none;
-    background: var(--gray-300, #c7c7c7);
-
+    border-radius: 100px;
+    background-color: #000;
     color: #fff;
-    cursor: pointer;
+}
+
+.homeLoginButton {
+    display: flex;
+    flex-direction: row;
+    color: #fff;
+    justify-content: space-between;
+    align-items: center;
+    gap: 14px;
 }
